Updated tokenizer
